# src/api-specifications/api-spec.yml
openapi: '3.0.3'
info:
  title: DaaS-NodeJS
  version: '1.1'
  
  contact:
    name: API Support
    email: alessio.farfaglia@gmail.com
  
  license:
    name: MPL-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: http://localhost:{port}/api
    description: Development server
    variables:
      port:
        default: '3000'
        description: default port for the development server
...
components:
  parameters:
    limit:
      name: limit
      in: query
      description: Numero massimo di oggetti da ritornare
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
      example: 20
    offset:
      name: offset
      in: query
      description: Offset per la paginazione
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
  responses:  
    InternalError:
      description: Errore interno del server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: integer
          description: ID del device
        device_model_id:
          type: integer
          description: ID del modello del device
        din_id:
          type: integer
          description: ID del DIN associato al device
        name:
          type: string
          description: Nome del device
          maxLength: 45
        latitude:
          type: number
          description: latitudine
          example: 39.3015
        longitude:
          type: number
          description: longitudine
          example: 16.2536
...
    DeviceModel:
      type: object
      properties:
        id:
          type: integer
          description: ID del modello del device
          nullable: false
        device_group_id:
          type: integer
          description: ID del gruppo del modello
          nullable: false
        description:
          type: string
          description: descrizione del modello
        serial:
          type: string
          description: seriale del modello
          maxLength: 20
    DeviceModelFull:
      type: object
      properties:
        id:
          type: integer
          description: ID del modello del device
          nullable: false
        device_group_id:
          type: integer
          description: ID del gruppo del modello
          nullable: false
        description:
          type: string
          description: descrizione del modello
        serial:
          type: string
          description: seriale del modello
          maxLength: 20
        device_group:
          $ref: '#/components/schemas/DeviceModelGroup'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/DeviceModelResource'
    DeviceModelGroup:
      type: object
      properties:
        id:
          type: integer
          description: ID del gruppo del modello
          nullable: false
        title:
          type: string
          description: titolo del gruppo
          maxLength: 45
        description:
          type: string
          description: descrizione del gruppo
        link_image:
          type: string
          format: uri
          description: collegamento all'immagine del gruppo
          example: "https://placehold.co/600x400"
    DeviceModelResource:
      type: object
      properties:
        id:
          type: integer
          description: ID della risorsa
          nullable: false
        name:
          type: string
          nullable: false
        device_model_id:
          type: integer
          description: ID del modello di device al quale si riferisce questa risorsa
          nullable: false
        link:
          type: string
          format: uri
          description: collegamento alla risorsa
          maxLength: 255
          example: "https://placehold.co/600x400"
        resource_type:
          type: integer
          enum:
            - 1
            - 2
            - 3
            - 4
            - 5
          description: >
            Tipo di risorsa

            1: altro

            2: scheda tecnica

            3: scheda utente

            4: immagine

            5: binario
          nullable: false
    DeviceModelFunctionProperty:
      description: >
        Property (param, input, output, notification) di una DeviceModelFunction
        Interpretabile come definizione per le properties di una DeviceFunction.
        
        Non contiene valori di istanza come 'value' che fanno invece parte di DeviceFunctionProperty
      type: object
      properties:
        id:
          type: integer
          description: ID della property
          nullable: false
        property_type:
          type: integer
          enum:
            - 1
            - 2
            - 3
            - 4
          description: >
            Tipo di property
            1: parameter
            2: input
            3: output
            4: notification
          nullable: false
        name:
          type: string
          description: nome della property
          maxLength: 45
        function_id:
          type: integer
          description: ID della DeviceModelFunction a cui si riferisce questa property
          nullable: false
        data_type:
          type: integer
          description: >
            tipo di dato
            1: i32
            2: i16
            3: f32
            4: string
          nullable: false
        default_value:
          type: string
          description: valore di default della property
          maxLength: 45
        safe_value:
          type: string
          description: valore sicuro della property, usato in caso di errore
          maxLength: 45
    DeviceModelFunction:
      type: object
      description: >
        Funzione di un modello di device
        Interpretabile come definizione per le funzioni di un device che sono istanze di questa funzione
      properties:
        id:
          type: integer
          description: ID della funzione
          nullable: false
        name:
          type: string
          description: nome della funzione
          maxLength: 45
        device_model_id:
          type: integer
          description: ID del modello di device a cui si riferisce questa funzione
          nullable: false
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/DeviceModelFunctionProperty'
          description: lista dei parametri della funzione
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/DeviceModelFunctionProperty'
          description: lista degli input della funzione
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/DeviceModelFunctionProperty'
          description: lista degli output della funzione
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/DeviceModelFunctionProperty'
          description: lista delle notifiche della funzione
...
    PaginationData: 
      description: Meta dati sulla paginazione dei dati
      properties:
        count:
          type: integer
          description: numero di oggetti contenuti tra i dati della risposta
          minimum: 0
          example: 20
        offset:
          type: integer
          description: gli oggetti contenuti dei dati della risposta partono da questo offset
          minimum: 40
        limit:
          type: integer
          description: numero massimo di oggetti contenuti tra i dati della risposta
          minimum: 0
          example: 20
        total:
          description: numero totale di oggetti per questa query
          type: integer
          minimum: 0
          example: 67
        has_next:
          type: boolean
          description: flag che indica se ci sono altri oggetti da visualizzare
          example: true
        has_prev:
          type: boolean
          description: flag che indica se ci sono oggetti precedenti da visualizzare
          example: false
    Error:
      type: object
      properties:
        error_name:
          type: string
          description: Nome dell'errore
          example: "TypeError"
        message:
          type: string
          description: Messaggio di errore

#region Paths
paths:
...
  /device_models:
    get:
      description: Ritorna una pagina di tutti i modelli di device.
      tags:
        - device models
      parameters:
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
        - name: q
          in: query
          description: >
            Filtro per nome del modello di device.
          schema:
            type: string
          example: "modello 2"

      responses:
        '200':
          description: Pagina di modelli di device e informazioni sulla paginazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceModelFull'
                  pagination:
                    $ref: '#/components/schemas/PaginationData'
                  q:
                    type: string
                    description: Filtro applicato
                    nullable: true

        default:
          $ref: '#/components/responses/InternalError'

    post:
      description: >
        Crea un nuovo modello di device, associando un gruppo di device esistente o creandone uno nuovo se non esiste.

        Permette di creare risorse di device da allegare al modello e DeviceModelFunctions per definire le funzioni che questo modello può eseguire.
      tags:
        - device models
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device_group_id:
                  type: integer
                  description: ID del gruppo del modello
                description:
                  type: string
                  description: Descrizione del modello
                serial:
                  type: string
                  description: Seriale del modello
                  maxLength: 20
                
                resources:
                  type: array
                  items:
                    $ref: '#/components/schemas/DeviceModelResource'
                    example:
                      name: "cover"
                      link: "https://m.media-amazon.com/images/I/71D4E5DS-qL.jpg"
                      resource_type: 1

                functions:
                  type: array
                  items:
                    $ref: '#/components/schemas/DeviceModelFunction'


              required:
                - device_group_id
                - description
                - serial
            example:
              description: ESP32
              serial: '1234567890'
              device_group:
                title: ESP32 DIY Modules
                description: ESP32 DIY Modules
                link_image: https://ce8dc832c.cloudimg.io/v7/_cdn_/BE/B6/90/00/0/617451_1.jpg?width=640&height=480&wat=1&wat_url=_tme-wrk_%2Ftme_new.png&wat_scale=100p&ci_sign=ecbc082e1968d44612ae5635e6defb9c957a3da9
              resources:
              - name: cover
                link: https://m.media-amazon.com/images/I/71D4E5DS-qL.jpg
                resource_type: 1
              - name: datasheet
                link: https://www.espressif.com/sites/default/files/documentation/esp32_datasheet_en.pdf
                resource_type: 2
              - name: hardware design guidelines
                link: https://www.espressif.com/sites/default/files/documentation/esp32_hardware_design_guidelines_en.pdf
                resource_type: 2
              - name: programming guide
                link: https://docs.espressif.com/projects/esp-idf/en/latest/esp32/get-started/index.html
                resource_type: 1
              functions:
              - name: PID1
                parameters:
                - name: P
                  data_type: 1
                - name: I
                  data_type: 1
                - name: D
                  data_type: 1
                - name: Frequenza
                  data_type: 1
                - name: Timeout
                  data_type: 1
                inputs:
                - name: target
                  data_type: 1
                - name: corrente
                  data_type: 1
                outputs:
                - name: pilota
                  data_type: 1
              - name: PLN1
                parameters:
                - name: ix
                  data_type: 1
                - name: week-day
                  data_type: 1
                - name: time
                  data_type: 1
                - name: pilota
                  data_type: 1
                outputs:
                - name: pilota
                  data_type: 1

      responses:
        '200':
          description: Modello di device aggiunto con successo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceModel'
        '400':
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
  /device_models/{deviceModelId}:
    parameters:
      - name: deviceModelId
        in: path
        description: ID del modello di device
        required: true
        schema:
          type: integer

    get:
      description: Ritorna un modello di device specifico.
      tags:
        - device models
      responses:
        '200':
          description: Modello di device trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceModelFull'
        '404':
          description: Modello di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'

    put:
      description: Aggiorna un modello di device specifico.
      tags:
        - device models
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device_group_id:
                  type: integer
                  description: ID del gruppo del modello
                description:
                  type: string
                  description: Descrizione del modello
                serial:
                  type: string
                  description: Seriale del modello
                  maxLength: 20
              example:
                device_group_id: 1
                description: "Modello di device 1"
                serial: "1234567890"
      responses:
        '200':
          description: Modello di device aggiornato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        '400':
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Modello di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
    delete:
      description: Elimina un modello di device specifico.
      tags:
        - device models
      responses:
        '200':
          description: Modello di device eliminato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        '404':
          description: Modello di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
  /device_models/{deviceModelId}/devices:
    parameters:
      - name: deviceModelId
        in: path
        description: ID del modello di device
        required: true
        schema:
          type: integer
    get:
      description: Ritorna i devices associati al modello di device specificato
      tags:
        - device models
        - devices
      parameters:
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
        - name: q
          in: query
          description: >
            Filtro per nome del device.
          schema:
            type: string
          example: "Device 1"
      responses:
        '200':
          description: Pagina di devices e informazioni sulla paginazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  pagination:
                    $ref: '#/components/schemas/PaginationData'
                  q:
                    type: string
                    description: Filtro applicato
                    nullable: true
        '404':
          description: Modello di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'        
        default:
          $ref: '#/components/responses/InternalError'
  /device_models/{deviceModelId}/resources:
    parameters:
      - name: deviceModelId
        in: path
        description: ID del modello di device
        required: true
        schema:
          type: integer
    get:
      description: Ritorna le risorse associate al modello di device specificato
      tags:
        - device models
      parameters:
        - name: type
          in: query
          description: >
            Filtro per tipo di risorsa.
          schema:
            type: integer
            enum:
              - 1
              - 2
              - 3
              - 4
              - 5
          example: 4
      responses:
        '200':
          description: Lista di risorse associate al modello di device.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceModelResource'
        '400':
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
    
    post:
      description: >
        Aggiunge una risorsa al modello di device.
      tags:
        - device models
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nome della risorsa
                  maxLength: 45
                link:
                  type: string
                  format: uri
                  description: collegamento alla risorsa
                  maxLength: 255
                  example: "https://placehold.co/600x400"
                resource_type:
                  type: integer
                  enum:
                    - 1
                    - 2
                    - 3
                    - 4
                    - 5
                  description: >
                    Tipo di risorsa

                    1: altro

                    2: scheda tecnica

                    3: scheda utente

                    4: immagine

                    5: binario
                  nullable: false
                  example: 1
              required:
                - name
                - link
                - resource_type
              example:
                name: "copertina"
                link: "https://placehold.co/200x300"
                resource_type: 4
      responses:
        '200':
          description: Risorsa aggiunta con successo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceModelResource'
        '400':
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Modello di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
  /device_models/{deviceModelId}/resources/{resourceId}:
    parameters:
      - name: deviceModelId
        in: path
        description: ID del modello di device
        required: true
        schema:
          type: integer
      - name: resourceId
        in: path
        description: ID della risorsa
        required: true
        schema:
          type: integer
    delete:
      description: Elimina una risorsa specifica associata al modello di device.
      tags:
        - device models
      responses:
        '200':
          description: Risorsa eliminata con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        '404':
          description: Modello di device o risorsa non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
      

#region Device Model Groups
...